// App/src/app/services/authentication/authentication-client.generated.ts
export class Credentials implements ICredentials {
    email?: string | undefined;
    password?: string | undefined;
    adminId?: string | undefined;
    applicationId?: string | undefined;

    constructor(data?: ICredentials) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.email = _data["email"];
            this.password = _data["password"];
            this.adminId = _data["adminId"];
            this.applicationId = _data["applicationId"];
        }
    }

    static fromJS(data: any): Credentials {
        data = typeof data === 'object' ? data : {};
        let result = new Credentials();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["email"] = this.email;
        data["password"] = this.password;
        data["adminId"] = this.adminId;
        data["applicationId"] = this.applicationId;
        return data; 
    }
}

export interface ICredentials {
    email?: string | undefined;
    password?: string | undefined;
    adminId?: string | undefined;
    applicationId?: string | undefined;
}

export class Ticket implements ITicket {
    registeredJWT?: string | undefined;
    registeredRefreshToken?: string | undefined;
    jwtIssuedAt?: string | undefined;
    applicationId?: string | undefined;

    constructor(data?: ITicket) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.registeredJWT = _data["registeredJWT"];
            this.registeredRefreshToken = _data["registeredRefreshToken"];
            this.jwtIssuedAt = _data["jwtIssuedAt"];
            this.applicationId = _data["applicationId"];
        }
    }

    static fromJS(data: any): Ticket {
        data = typeof data === 'object' ? data : {};
        let result = new Ticket();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["registeredJWT"] = this.registeredJWT;
        data["registeredRefreshToken"] = this.registeredRefreshToken;
        data["jwtIssuedAt"] = this.jwtIssuedAt;
        data["applicationId"] = this.applicationId;
        return data; 
    }
}
// ...
export class AccountRegistration implements IAccountRegistration {
    email?: string | undefined;
    configData?: string | undefined;
    password?: string | undefined;
    authenticationRole?: string | undefined;
    applicationId?: string | undefined;
    adminId?: string | undefined;

    constructor(data?: IAccountRegistration) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.email = _data["email"];
            this.configData = _data["configData"];
            this.password = _data["password"];
            this.authenticationRole = _data["authenticationRole"];
            this.applicationId = _data["applicationId"];
            this.adminId = _data["adminId"];
        }
    }

    static fromJS(data: any): AccountRegistration {
        data = typeof data === 'object' ? data : {};
        let result = new AccountRegistration();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["email"] = this.email;
        data["configData"] = this.configData;
        data["password"] = this.password;
        data["authenticationRole"] = this.authenticationRole;
        data["applicationId"] = this.applicationId;
        data["adminId"] = this.adminId;
        return data; 
    }
}

export interface IAccountRegistration {
    email?: string | undefined;
    configData?: string | undefined;
    password?: string | undefined;
    authenticationRole?: string | undefined;
    applicationId?: string | undefined;
    adminId?: string | undefined;
}

export class AccountConfirmation implements IAccountConfirmation {
    id?: string | undefined;
    message?: string | undefined;

    constructor(data?: IAccountConfirmation) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.message = _data["message"];
        }
    }

    static fromJS(data: any): AccountConfirmation {
        data = typeof data === 'object' ? data : {};
        let result = new AccountConfirmation();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["message"] = this.message;
        return data; 
    }
}

export interface IAccountConfirmation {
    id?: string | undefined;
    message?: string | undefined;
}
// ...
export class DomainName implements IDomainName {
    name?: string | undefined;
    url?: string | undefined;

    constructor(data?: IDomainName) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.name = _data["name"];
            this.url = _data["url"];
        }
    }

    static fromJS(data: any): DomainName {
        data = typeof data === 'object' ? data : {};
        let result = new DomainName();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["name"] = this.name;
        data["url"] = this.url;
        return data; 
    }
}

export interface IDomainName {
    name?: string | undefined;
    url?: string | undefined;
}

export enum ClaimType {
    _0 = 0,
    _1 = 1,
    _2 = 2,
    _3 = 3,
    _4 = 4,
    _5 = 5,
}

export class ClaimConfig implements IClaimConfig {
    jwtName!: string;
    type!: string;
    data?: string | undefined;
    claimConfigurations?: ClaimConfig[] | undefined;

    constructor(data?: IClaimConfig) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.jwtName = _data["jwtName"];
            this.type = _data["type"];
            this.data = _data["data"];
            if (Array.isArray(_data["claimConfigurations"])) {
                this.claimConfigurations = [] as any;
                for (let item of _data["claimConfigurations"])
                    this.claimConfigurations!.push(ClaimConfig.fromJS(item));
            }
        }
    }

    static fromJS(data: any): ClaimConfig {
        data = typeof data === 'object' ? data : {};
        let result = new ClaimConfig();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["jwtName"] = this.jwtName;
        data["type"] = this.type;
        data["data"] = this.data;
        if (Array.isArray(this.claimConfigurations)) {
            data["claimConfigurations"] = [];
            for (let item of this.claimConfigurations)
                data["claimConfigurations"].push(item.toJSON());
        }
        return data; 
    }
}

export interface IClaimConfig {
    jwtName: string;
    type: string;
    data?: string | undefined;
    claimConfigurations?: ClaimConfig[] | undefined;
}

export class JwtTenantConfig implements IJwtTenantConfig {
    secretKey!: string;
    claims!: ClaimConfig[];
    expireMinutes?: number | undefined;
    refreshExpireMinutes?: number | undefined;
    lockoutEnabled!: boolean;
    validateIssuer!: boolean;
    issuer?: string | undefined;
    algorithm!: string;

    constructor(data?: IJwtTenantConfig) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        if (!data) {
            this.claims = [];
        }
    }

    init(_data?: any) {
        if (_data) {
            this.secretKey = _data["secretKey"];
            if (Array.isArray(_data["claims"])) {
                this.claims = [] as any;
                for (let item of _data["claims"])
                    this.claims!.push(ClaimConfig.fromJS(item));
            }
            this.expireMinutes = _data["expireMinutes"];
            this.refreshExpireMinutes = _data["refreshExpireMinutes"];
            this.lockoutEnabled = _data["lockoutEnabled"];
            this.validateIssuer = _data["validateIssuer"];
            this.issuer = _data["issuer"];
            this.algorithm = _data["algorithm"];
        }
    }

    static fromJS(data: any): JwtTenantConfig {
        data = typeof data === 'object' ? data : {};
        let result = new JwtTenantConfig();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["secretKey"] = this.secretKey;
        if (Array.isArray(this.claims)) {
            data["claims"] = [];
            for (let item of this.claims)
                data["claims"].push(item.toJSON());
        }
        data["expireMinutes"] = this.expireMinutes;
        data["refreshExpireMinutes"] = this.refreshExpireMinutes;
        data["lockoutEnabled"] = this.lockoutEnabled;
        data["validateIssuer"] = this.validateIssuer;
        data["issuer"] = this.issuer;
        data["algorithm"] = this.algorithm;
        return data; 
    }
}
// ...
export class Application implements IApplication {
    id?: string | undefined;
    name?: string | undefined;
    domains?: DomainName[] | undefined;
    iconUUID?: string | undefined;
    jwtTenantConfigurations?: JwtTenantConfig[] | undefined;

    constructor(data?: IApplication) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            if (Array.isArray(_data["domains"])) {
                this.domains = [] as any;
                for (let item of _data["domains"])
                    this.domains!.push(DomainName.fromJS(item));
            }
            this.iconUUID = _data["iconUUID"];
            if (Array.isArray(_data["jwtTenantConfigurations"])) {
                this.jwtTenantConfigurations = [] as any;
                for (let item of _data["jwtTenantConfigurations"])
                    this.jwtTenantConfigurations!.push(JwtTenantConfig.fromJS(item));
            }
        }
    }

    static fromJS(data: any): Application {
        data = typeof data === 'object' ? data : {};
        let result = new Application();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        if (Array.isArray(this.domains)) {
            data["domains"] = [];
            for (let item of this.domains)
                data["domains"].push(item.toJSON());
        }
        data["iconUUID"] = this.iconUUID;
        if (Array.isArray(this.jwtTenantConfigurations)) {
            data["jwtTenantConfigurations"] = [];
            for (let item of this.jwtTenantConfigurations)
                data["jwtTenantConfigurations"].push(item.toJSON());
        }
        return data; 
    }
}